Removed bitmap, migrated opencv stuff from python to c++, added ring buffer

+++ src/ffi/constants.rs
@@ -2,8 +2,8 @@
 /// so that rust and c both agree on the values no matter what
 use crate::ffi::bindings;
 
-pub const WIDTH: usize = bindings::WIDTH;
-pub const HEIGHT: usize = bindings::HEIGHT;
-pub const SIZE: usize = bindings::SIZE;
+pub const WIDTH: usize = bindings::WIDTH as usize;
+pub const HEIGHT: usize = bindings::HEIGHT as usize;
+pub const SIZE: usize = bindings::SIZE as usize;
 pub const BACKGROUND: u8 = bindings::BACKGROUND;
 pub const FOREGROUND: u8 = bindings::FOREGROUND;
